backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.13.1
python-dotenv>=1.0.0
openai>=1.57.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
mcp>=1.1.2
aiosqlite>=0.19.0
httpx>=0.28.0
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
nest_asyncio>=1.6.0